requirements.txt
streamlit==1.22.0
torch==2.0.1
Pillow==9.5.0  # For image processing
numpy==1.25.0
